Build an enhancing 'deferrantize' to up-vert anything into a thenable.
Decouple the basic class from a larger, bigger helper that does most of the
work, turning anything into a thennable. Note that when I say thennable, I
mean an object with the Promise shape, offering .then, .catch, .finally.

This upgrade/enhancement will build .catch/.finally from an underlying
.then if .then is available.

Keep in mind, in addition, a bunch of state variables are exposed on the
deferrant or deferrantized object: resolved, rejected, fulfilled.